Popup et séparation de main.py en deux fichiers: main et application
